requirements.txt
backcall==0.2.0
certifi==2021.10.8
gunicorn==20.1.0
cffi==1.15.0
colorama==0.4.4
Flask==2.0.2
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipykernel==6.6.0
Jinja2==3.0.3
joblib==1.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
nbconvert==6.3.0
numpy==1.19.5
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
tensorflow-cpu==2.5.0
termcolor==1.1.0
terminado==0.12.1
webencodings==0.5.1
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.6.0
